+++ GeoAdServer/GeoAdServer.WebApi/Controllers/OfferingsController.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Web.Http;
+using GeoAdServer.WebApi.Support;
 
 namespace GeoAdServer.WebApi.Controllers
 {
@@ -22,21 +23,36 @@ namespace GeoAdServer.WebApi.Controllers
             return DataRepos.Offerings.GetByLocationId(id).AsQueryable();
         }
 
+        [Authorize]
         public HttpResponseMessage Post(Offering offering)
         {
+            if (offering == null) return Request.CreateResponse(HttpStatusCode.BadRequest);
+
+            if (!offering.LocationId.IsLocationOwner(RequestContext.GetUserId()))
+                return Request.CreateResponse(HttpStatusCode.Unauthorized);
+
             int id = DataRepos.Offerings.Insert(offering);
             return id != -1 ? Request.CreateResponse(HttpStatusCode.OK, id) :
                               Request.CreateResponse(HttpStatusCode.InternalServerError);
         }
 
+        [Authorize]
         public HttpResponseMessage Put(int id, [FromBody]Offering offering)
         {
+            if (offering == null) return Request.CreateResponse(HttpStatusCode.BadRequest);
+
+            if (!id.IsLocationOwner(RequestContext.GetUserId()))
+                return Request.CreateResponse(HttpStatusCode.Unauthorized);
+
             var result = DataRepos.Offerings.Update(id, offering);
             return Request.CreateResponse(result ? HttpStatusCode.NoContent : HttpStatusCode.NotFound);
         }
 
         public HttpResponseMessage Delete(int id)
         {
+            if (!id.IsLocationOwner(RequestContext.GetUserId()))
+                return Request.CreateResponse(HttpStatusCode.Unauthorized);
+
             foreach (OfferingDTO off in DataRepos.Offerings.GetByLocationId(id))
             {
                 DataRepos.Offerings.DeleteById(off.Id);
